Small changes, see description
Add compile fail test to ensure the drop check for SendError is working
correctly

Fix comment formatting in SendError::new

Add PhantomData<T> to Sender and Receiver. This may not be strictly
necessary, but is not incorrect (whereas not adding it might be).

Bump MSRV to 1.60.0 for MaybeUninit::assume_init_drop